Download energy data too

+++ tesla_solar_download.py
@@ -98,7 +98,7 @@ def _download_data(tesla, site_id, debug=False):
         if partial_day or not os.path.exists(_get_csv_name(date, site_id)):
             print(date.isoformat())
             _download_day(tesla, site_id, timezone, date, partial_day=partial_day)
-            time.sleep(3)
+            time.sleep(1)
         date -= timedelta(days=1)
         partial_day = False
         # Re-localize the date based on the timezone.  This is important because we maybe have
@@ -106,8 +106,82 @@ def _download_data(tesla, site_id, debug=False):
         date = pytz.timezone(timezone).localize(date.replace(tzinfo=None))
 
 
+def _get_energy_csv_name(site_id):
+    return f'download/{site_id}/energy.csv'
+
+
+def _write_energy_csv(timeseries, site_id):
+    if not timeseries:
+        raise ValueError('No timeseries')
+
+    csv_filename = _get_energy_csv_name(site_id)
+    os.makedirs(os.path.dirname(csv_filename), exist_ok=True)
+    file_exists = os.path.exists(csv_filename)
+    fieldnames = list(timeseries[0].keys())
+    with open(csv_filename, 'a') as csv_file:
+        writer = csv.DictWriter(csv_file, fieldnames=fieldnames)
+        if not file_exists:
+            writer.writeheader()
+        for ts in timeseries:
+            ts['timestamp'] = parse(ts['timestamp']).strftime('%Y-%m-%d %H:%M:%S')
+            writer.writerow(ts)
+
+
+@retry(tries=2, delay=5)
+def _download_energy_month(tesla, site_id, timezone, start_date, end_date):
+    print(start_date.isoformat(), end_date.isoformat())
+    response = tesla.api(
+        'CALENDAR_HISTORY_DATA',
+        path_vars={'site_id': site_id},
+        kind='energy',
+        period='month',
+        start_date=start_date.isoformat(),
+        end_date=end_date.isoformat(),
+        time_zone=timezone,
+        fill_telemetry=0,
+    )['response']
+
+    return response['time_series']
+
+
+def _download_energy_data(tesla, site_id, debug=False):
+    site_config = tesla.api('SITE_CONFIG', path_vars={'site_id': site_id})['response']
+    installation_date = parse(site_config['installation_date'])
+    timezone = site_config['installation_time_zone']
+
+    now = datetime.now(pytz.timezone(timezone)).replace(microsecond=0)
+    start_date = now.replace(hour=0, minute=0, second=0)
+    end_date = now.replace(hour=23, minute=59, second=59)
+
+    csv_name = _get_energy_csv_name(site_id)
+    if os.path.exists(csv_name):
+        os.remove(csv_name)
+    # Beginning of the month.
+    start_date = start_date - timedelta(days=start_date.day - 1)
+    if debug:
+        print(f'Timezone: {timezone}')
+        print(f'Start date: {start_date}')
+
+    timeseries = []
+    while end_date > installation_date:
+        timeseries.extend(
+            _download_energy_month(tesla, site_id, timezone, start_date, end_date)
+        )
+        time.sleep(1)
+        end_date = start_date - timedelta(seconds=1)
+        start_date = end_date.replace(hour=0, minute=0, second=0) - timedelta(
+            days=end_date.day - 1
+        )
+        start_date = pytz.timezone(timezone).localize(start_date.replace(tzinfo=None))
+
+    timeseries.sort(key=lambda x: x['timestamp'])
+    _write_energy_csv(timeseries, site_id)
+
+
 def _delete_partial_files(site_id):
     dir = os.path.join('download', str(site_id))
+    if not os.path.exists(dir):
+        return
     for fname in os.listdir(dir):
         if '.partial.csv' in fname:
             os.remove(os.path.join(dir, fname))
@@ -142,6 +216,7 @@ def main():
             print(f'Downloading data for {resource_type} site ***{str(site_id)[-4:]}')
             _delete_partial_files(site_id)
             _download_data(tesla, site_id, debug=args.debug)
+            _download_energy_data(tesla, site_id, debug=args.debug)
 
 
 if __name__ == '__main__':
